Changed: Inlined the builder into the main determine_optimal_transform
