+++ projects/api/dxt-lossless-transform-bc1-api/src/c_api/mod.rs
@@ -3,6 +3,6 @@
 //! This module provides C-compatible exports for BC1 texture compression transforms.
 //! All functions are prefixed with `dltbc1_` for uniqueness.
 
-pub mod determine_optimal_transform;
+mod determine_optimal_transform;
 pub mod error;
 pub mod transform;
